requirements.txt
streamlit
torch
torchvision
transformers
accelerate
bitsandbytes
Pillow
numpy
scipy
matplotlib
seaborn
deepspeed
